+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### 3.17.1(53)
+
+Network Client
+
+- `NetworkError` is renamed to `NetworkClientError`
+
 ### [3.17.0(52)](https://github.com/VakhoKontridze/VCore/releases/tag/3.17.0) — *2022 07 23*
 
 Services and Managers